Fixing test single
