+++ requirements.txt
@@ -7,3 +7,4 @@ wandb==0.10.20
 mypy==0.812
 hydra-core==1.0.6
 omegaconf==2.0.6
+code2seq==0.0.3
